requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.42.0
python-dotenv==1.0.1
httpx==0.27.0
